CareerMatch/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
beautifulsoup4==4.12.2
supabase==2.3.5
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0 